requirements.txt
# For run the app ------> python3 -m streamlit run app.py

# Core Framework
streamlit>=1.28.0
streamlit-extras>=0.3.0

# Data Fetching
yfinance>=0.2.30
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Machine Learning - Use LightGBM instead of XGBoost
lightgbm>=4.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
xgboost

# NLP & Sentiment Analysis
transformers>=4.30.0
torch>=2.0.0

# Technical Analysis
ta>=0.10.0

# Visualization
plotly>=5.0.0
matplotlib>=3.7.0

# Explainability
shap>=0.42.0

# PDF Generation
reportlab>=4.0.0

# News APIs
gnews>=0.3.0

# Utilities
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pytz>=2024.0
